pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "learning"
authors = [
    {name = "zxc-anton",email = "antontaradaev8@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "flask-moment (>=1.0.6,<2.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "flask-babel (>=4.0.0,<5.0.0)",
    "elasticsearch (>=9.0.0,<10.0.0)",
    "rq (>=2.3.2,<3.0.0)",
    "flask-httpauth (>=4.8.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flask-migrate = "^4.1.0"
aiosmtpd = "^1.4.6"
httpie = "^3.2.4"

